Extract ApiProblem creation to a separate class
